Test with Secret Config File
